requirements.txt
pydantic-ai>=0.1.0
crawl4ai>=0.1.0
supabase>=2.0.0
openai>=1.0.0
python-dotenv>=0.19.0
httpx>=0.24.0
logfire>=0.8.0
streamlit>=1.30.0
PyPDF2>=3.0.0
requests>=2.31.0
asyncio>=3.4.3
